radar/radar copy/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-bootstrap-form==3.2
django-bootstrap-pagination==1.5.1
django-import-export==0.2.8
django-toolbelt==0.0.1
gevent==1.0.2
greenlet==0.4.7
grequests==0.2.0
gunicorn==19.3.0
kombu==3.0.26
MySQL-python==1.2.5
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
tablib==0.10.0
wheel==0.24.0
